github: Add workaround for running tests in the container
User in the container will be root, but the cloned repo is owned by the
user running github-runner. Git doesn't like that as the repo will then
be unusable by the owner:  we don't care about it since this is only
used for running these tests.

See https://github.com/actions/runner/issues/2033

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>