genbranch: Fix cache update for non-inplace case
Use the correct head revision for the cache update. Without this fix,
any update applied to the cache when running genbranch without the `-i`
option makes wrong equivalence association between patches and commits.

Because of already affected users, also bump the required value for the
schema version, even though the schema remains the same: that will make
the corrupted cache to be dismissed in favor of an empty one.

Fixes: 23ba85f1ef41 ("genbranch: Support caching")
Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>