cli: Do config handling inside PileCLI
A new option --no-config will be added which would make config handling
depending on that. As such, it makes sense to move the current logic
inside PileCLI, since global pile CLI options are defined in that class.

Note that we do not need to check for args.command being None before
calling config.normalize(), because at that point we would have already
returned if that was true (see the try/except block above).

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>