git-merge-range: Add command
This is a useful command to perform merges of revision ranges using a
strategy based on the 3-way merge of files.

In multiple situations it was proven very useful to re-apply a pile
series, avoiding conflicts that would arise otherwise. In an upcoming
change, a new strategy for `git pile am` will be added to allow using
this, where details of how this is used for pile will be provided.

Note that this is kept as a separate "self-contained" directory because
this can be a separate command of its own, independent of git-pile. In
the future we might try incorporating it into git.

The current implementation is quite lazy and uses git-range-diff to find
matching commits between ranges instead of implementing its own logic.
As we are currently only interested in range-diff lines with "=" type of
matches, we could get away with an algorithm not so time-complex as the
one used by git-range-diff. We can do such optimization in the future.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>