am: Introduce the merge-range strategy
It is not uncommon to have conflicts when applying a large series to the
pile (like a rebase onto a new baseline) after some development from
other contributors has been incorporated since the creation of the
series.

Using git-merge-range has proven to be very useful in those situations,
because it reduces the number of conflicts that one would run into when
applying the series and, if conflicts happen, they are arguably easier
to work with, since we handle them at the "file level" instead of
working directly on the patches.

While one can do the steps to use git-merge-range manually, it is much
better to the user that the "am" operation automates that by having a
"merge-range" strategy, so let's implement it.

It would be nice to have all tests in 60_format_patch_and_am.bats to be
parametrized by the strategy, so that we have more coverage. However, an
attempt to use bats_test_function as in [1] failed with bats 1.10.0.
That's why, for now, we fork only test format-patch-and-am-strategy, so
that there is one for the "top" strategy and another for the
"merge-range" strategy.

[1] https://github.com/bats-core/bats-core/blob/990d8e285e7564ff61ea2c93a6da12856d337d83/test/fixtures/bats/dynamic_test_registration.bats

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>