gitutil: Try to remove temporary workdir first
We the current state of git_temporary_worktree(), probably due to
timing, sometimes we would end up with a failure when removing the
worktree, most likely because the cleanup performed TemporaryDirectory
took place in the file system.

Let's invert stuff here by having TemporaryDirectory wrapping the try
statement so that we gracefully remove the worktree before the cleanup
operation kicks in.

Note that with this change, we also need to use the --force flag to the
worktree removal command, as there might be unstaged/untracked files in
it.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>